refactor: extract DB connection setup and utilities into src/backend/db.ts
Move initDB(), PadString(), stableHash(), generateSeed(), getSqliteMaxVariables(),
and computeBatchSize() out of backend.ts into a new focused db.ts module to slim
down the 2829-line backend.ts as part of the codebase manageability refactor.
